fix(skills): move command-exists to core layer and declare skills config surface
